move source to the separate file
